LineSymbol: Fix border line handling
The path of a border line is determined by two factors: the main
offset, indicating the edge of the main line, and an explicit
additional offset.
Mapper used to rely on the main offset being non-zero in order to
distinguish left from right by the sign of the main offset.
With this change, Mapper now uses the totat offset to determine left
or right, thus allowing to use border lines with invisible main lines.
Fixes GH-1602 (wrong area border line path).